Create funcationalarry.c
A student wants to initialize and print a 2-dimensional integer array in a program. However, he/she is forbidden to declare 2-dimensional arrays in the main function of the program. Assume that the 2-dimensional integer array is of size 4 rows and 4 columns where n=16 is the total number of elements of the 1D array. 

You are required to write the following two functions.

Function 1:  'int * getArray1D (int n)': The function should return pointer to 1D array in the function 'getArray2D' from where it is called [5 Marks]

Function 2: 'void convertArray2D (int n) ': The function should convert 1D array to 2D and print it. It should call the above function 'getArray1D' to get 1D array of size 'n'. [5 Marks]

The code of main function is shown below:
int main () {
  	int n=16;
  	convertArray2D (n);
    return 0;
}